requirements.txt
# RCM Schema Dependencies
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pgvector>=0.2.5
pydantic>=2.5.0
python-dotenv>=1.0.0